+++ metrics/prometheus/promhttp/module.go
@@ -1,0 +1,9 @@
+package fxpromhttp
+
+import "go.uber.org/fx"
+
+// Module is an fx compatible module.
+var Module = fx.Provide(
+	NewHandler,
+	NewLogger,
+)